requirements.txt
wheel
requests
sqlalchemy
python-telegram-bot
bitmex
websocket
bitmex-ws
asyncio
dependency_injector
